# perf_capture_timer preload resources to reduce queries