--- a/app/models/zone.rb
+++ b/app/models/zone.rb
@@ -161,10 +161,9 @@ class Zone < ActiveRecord::Base
   end
 
   def vms_without_availability_zone
-    MiqPreloader.preload(self, :ext_management_systems => :vms)
-    ext_management_systems.flat_map do |e|
-      e.kind_of?(EmsCloud) ? e.vms.select { |vm| vm.availability_zone.nil? } : []
-    end
+    clouds = ext_management_systems.select { |e| e.kind_of?(EmsCloud) }
+    MiqPreloader.preload(clouds, :vms => :availability_zone)
+    clouds.flat_map { |e| e.vms.select { |vm| vm.availability_zone.nil? } }
   end
 
   def vms_and_templates