We have many queries for perf_capture_timer. This aims to reduce them.

Still have one more pass after this

- We were looking up `Zone` multiple times.
- We were loading all the vms without an availability zone even when not necessary.
- Made a number of smaller changes, like using `flat_map` and `with_object`, which will have a negligible effect.
- Split up the primary method so we could focus on the slower portion of code.

||before|after
---|---|---
vms|10k|10k|
miq_queue|7950|7950|
miq_task|1543|7950|
time|6750ms|4729 ms
sql count|879|679
sql time|323ms|191ms
memory|190mb|110mb

**DISCLAIMER:** this does not include `capture_queue` (which is the biggest method) (it was timing out on my box)

https://bugzilla.redhat.com/show_bug.cgi?id=1221750