@kbrock 2016-01-12
@Fryguy that first suggestion was beautiful - will check second one (think it will be worse though)

@jrafanie thanks for keeping me honest. document all things!