@Fryguy 2016-01-12
This looks great @kbrock.  Let's get it in as is without any more changes.  These are very atomic and solve the problem...not need to clutter up the finished good decisions with more changes...you can always make a second PR.

I made a minor comment about making those new methods private.  Once that's in, this is good to go.