@kbrock 2016-01-12
@Fryguy updated

- I moved the code around to put the privates together. Was tricky to make the class methods private.
- Fun how the methods I extracted are too complex. What about the original? ;)
- Fixed specs.